Finished disentangling Opentxs from the CLI.
