+++ include/opentxs/client/OTRecordList.hpp
@@ -166,6 +166,9 @@ typedef std::map<std::string, std::string> map_of_strings;
 
 class OTRecordList
 {
+public:
+    enum ItemType { typeBoth = 0, typeTransfers = 1, typeReceipts = 2 };
+
 private:
     const OTNameLookup* m_pLookup{nullptr};
     // Defaults to false. If you set it true, it will run a lot faster. (And
@@ -244,6 +247,32 @@ public:
         const std::string& mynym,
         const std::string& instrument,
         std::string* pOptionalOutput =nullptr);
+
+    EXPORT static std::int32_t acceptFromInbox(
+        const std::string& myacct,
+        const std::string& indices,
+        const std::int32_t itemTypeFilter);
+
+    EXPORT static bool checkMandatory(const char* name,
+                                      const std::string& value);
+
+    EXPORT static bool checkNym(const char* name,
+                                std::string& nym,
+                                bool checkExistance=true);
+    EXPORT static bool checkIndices(const char* name, const std::string& indices);
+    EXPORT static bool checkAccount(const char* name, std::string& account);
+
+    EXPORT static bool checkServer(const char* name, std::string& server);
+
+    EXPORT static std::int32_t discard_incoming_payments(
+        const std::string& server,
+        const std::string& mynym,
+        const std::string& indices);
+    
+    EXPORT static std::int32_t cancel_outgoing_payments(
+        const std::string& mynym,
+        const std::string& myacct,
+        const std::string& indices);
 
     EXPORT static std::string inputText(const char* what);
