+++ src/client/OTRecordList.cpp
@@ -58,6 +58,7 @@
 #include "opentxs/contact/ContactData.hpp"
 #include "opentxs/core/contract/UnitDefinition.hpp"
 #include "opentxs/core/recurring/OTPaymentPlan.hpp"
+#include "opentxs/core/script/OTSmartContract.hpp"
 #include "opentxs/core/util/Assert.hpp"
 #include "opentxs/core/util/Common.hpp"
 #include "opentxs/core/Account.hpp"
@@ -1176,6 +1177,697 @@ std::int32_t OTRecordList::confirmPaymentPlan_lowLevel( // a static method
 
     return 1;
 }
+
+//static
+bool OTRecordList::checkMandatory(const char* name,
+                                  const std::string& value)
+{
+    if (value.empty()) {
+        otOut << "Error: " << name << ": mandatory parameter not specified.\n";
+        return false;
+    }
+
+    return true;
+}
+
+//static
+bool OTRecordList::checkIndices(const char* name, const std::string& indices)
+{
+    if (!checkMandatory(name, indices)) {
+        return false;
+    }
+
+    if ("all" == indices) {
+        return true;
+    }
+
+    for (std::string::size_type i = 0; i < indices.length(); i++) {
+        if (!isdigit(indices[i])) {
+            otOut << "Error: " << name << ": not a value: " << indices << "\n";
+            return false;
+        }
+        for (i++; i < indices.length() && isdigit(indices[i]); i++) {
+        }
+        if (i < indices.length() && ',' != indices[i]) {
+            otOut << "Error: " << name << ": not a value: " << indices << "\n";
+            return false;
+        }
+    }
+
+    return true;
+}
+
+// static
+bool OTRecordList::checkServer(const char* name, std::string& server)
+{
+    if (!checkMandatory(name, server))
+        return false;
+
+    Identifier theID(server);
+    ConstServerContract pServer;  // shared_ptr to const.
+
+    // See if it's available using the full length ID.
+    if (!theID.empty())
+        pServer = OT::App().Wallet().Server(theID);
+
+    if (!pServer) {
+        const auto servers = OT::App().Wallet().ServerList();
+
+        // See if it's available using the partial length ID.
+        for (auto& it : servers) {
+            if (0 == it.first.compare(0, server.length(), server)) {
+                pServer = OT::App().Wallet().Server(Identifier(it.first));
+                break;
+            }
+        }
+        if (!pServer) {
+            // See if it's available using the full length name.
+            for (auto& it : servers) {
+                if (0 == it.second.compare(0, it.second.length(), server)) {
+                    pServer = OT::App().Wallet().Server(Identifier(it.first));
+                    break;
+                }
+            }
+
+            if (!pServer) {
+                // See if it's available using the partial name.
+                for (auto& it : servers) {
+                    if (0 == it.second.compare(0, server.length(), server)) {
+                        pServer =
+                            OT::App().Wallet().Server(Identifier(it.first));
+                        break;
+                    }
+                }
+            }
+        }
+    }
+
+    if (!pServer) {
+        otOut << "Error: " << name << ": unknown server: " << server << "\n";
+        return false;
+    }
+
+    server = pServer->ID().str();
+    otOut << "Using " << name << ": " << server << "\n";
+    return true;
+}
+
+
+// static
+bool OTRecordList::checkNym(const char* name, std::string& nym,
+                            bool checkExistance/*=false*/)
+{
+    if (!checkMandatory(name, nym))
+        return false;
+
+    ConstNym pNym = nullptr;
+    const Identifier nymID(nym);
+
+    if (!nymID.empty())
+        pNym = OT::App().Wallet().Nym(nymID);
+
+    if (nullptr == pNym)
+        pNym = OT::App().Wallet().NymByIDPartialMatch(nym);
+
+    if (nullptr != pNym) {
+        String tmp;
+        pNym->GetIdentifier(tmp);
+        nym = tmp.Get();
+    } else if (checkExistance) {
+        otOut << "Error: " << name << ": unknown nym: " << nym << "\n";
+        return false;
+    }
+
+    otOut << "Using " << name << ": " << nym << "\n";
+    return true;
+}
+
+//static
+bool OTRecordList::checkAccount(const char* name, std::string& account)
+{
+    if (!checkMandatory(name, account)) {
+        return false;
+    }
+
+    std::shared_ptr<Account> pAccount{nullptr};
+    OTWallet* wallet = OT::App().API().OTAPI().GetWallet();
+
+    Identifier theID(account);
+
+    if (!theID.empty())
+        pAccount = wallet->GetAccount(theID);
+
+    if (false == bool(pAccount)) {
+        pAccount = wallet->GetAccountPartialMatch(account);
+
+        if (false == bool(pAccount)) {
+            otOut << "Error: " << name << ": unknown account: " << account
+                  << "\n";
+            return false;
+        }
+    }
+
+    if (pAccount) {
+        String tmp;
+        pAccount->GetPurportedAccountID().GetString(tmp);
+        account = tmp.Get();
+    }
+
+    otWarn << "Using " << name << ": " << account << "\n";
+    return true;
+}
+
+//static
+std::int32_t OTRecordList::discard_incoming_payments(
+    const std::string& server,
+    const std::string& mynym,
+    const std::string& indices)
+{
+    std::string the_server = server;
+    if (!checkServer("server", the_server)) {
+        return -1;
+    }
+
+    std::string the_mynym = mynym;
+    if (!checkNym("mynym", the_mynym)) {
+        return -1;
+    }
+
+    if (!checkIndices("indices", indices)) {
+        return -1;
+    }
+
+    std::string inbox = SwigWrap::LoadPaymentInbox(the_server, the_mynym);
+    if (inbox.empty()) {
+        otOut << "Error: cannot load payment inbox.\n";
+        return -1;
+    }
+
+    std::int32_t items = SwigWrap::Ledger_GetCount(the_server,
+                                                   the_mynym, the_mynym, inbox);
+    if (0 > items) {
+        otOut << "Error: cannot load payment inbox item count.\n";
+        return -1;
+    }
+
+    if (0 == items) {
+        otOut << "The payment inbox is empty.\n";
+        return 0;
+    }
+
+    bool all = ("all" == indices);
+
+    // Loop from back to front, in case any are removed.
+    std::int32_t retVal = 1;
+    for (std::int32_t i = items - 1; 0 <= i; i--) {
+        if (!all && !SwigWrap::NumList_VerifyQuery(indices,
+                                                   std::to_string(i))) {
+            continue;
+        }
+
+        if (!SwigWrap::RecordPayment(the_server, the_mynym, true, i, false)) {
+            otOut << "Error: cannot discard payment.\n";
+            retVal = -1;
+            continue;
+        }
+
+        otOut << "Success discarding payment!\n";
+    }
+
+    return retVal;
+}
+
+//static
+std::int32_t OTRecordList::cancel_outgoing_payments(
+    const std::string& mynym,
+    const std::string& myacct,
+    const std::string& indices)
+{
+    // NOTE: You can't just discard a sent cheque  from your outpayment box.
+    // Why not? Just because you remove your record of the outgoing cheque,
+    // doesn't mean you didn't already send it. (The recipient still received
+    // it, and still has it, whether you remove it from your outbox or not.)
+    // If you really want to cancel the cheque, then you need to do it in such
+    // a way that it will fail-as-cancelled when the recipient tries to deposit
+    // it. Otherwise, he would get that money whether you kept your own outgoing
+    // record or not. Therefore SOME server message must be performed here,
+    // which actually cancels the transaction number itself that appears on the
+    // cheque. This is the only way to insure that the cheque can't be used by
+    // the recipient (and even this will only work if you beat him to the punch.
+    // If he deposits it before you cancel it, then it's already too late and he
+    // has the money.) THIS is why RecordPayment, regarding outpayments, only
+    // works on expired instruments -- because if it's not expired, you don't
+    // just want to remove your record of it. You want to cancel the transaction
+    // number itself -- and that requires server communication.
+
+    std::string the_mynym = mynym;
+    if (!checkNym("mynym", the_mynym)) {
+        return -1;
+    }
+
+    std::string the_myacct = myacct;
+    if (!myacct.empty() && !checkAccount("myacct", the_myacct)) {
+        return -1;
+    }
+
+    if (!checkIndices("indices", indices)) {
+        return -1;
+    }
+
+    std::int32_t items = SwigWrap::GetNym_OutpaymentsCount(the_mynym);
+    if (0 > items) {
+        otOut << "Error: cannot load payment outbox item count.\n";
+        return -1;
+    }
+
+    if (0 == items) {
+        otOut << "The payment outbox is empty.\n";
+        return 0;
+    }
+
+    bool all = ("all" == indices);
+
+    // Loop from back to front, in case any are removed.
+    std::int32_t retVal = 1;
+    for (std::int32_t i = items - 1; 0 <= i; i--) {
+        if (!all && !SwigWrap::NumList_VerifyQuery(indices,
+                                                   std::to_string(i))) {
+            continue;
+        }
+
+        std::string payment = SwigWrap::GetNym_OutpaymentsContentsByIndex(the_mynym, i);
+        if (payment.empty()) {
+            otOut << "Error: cannot load payment " << i << ".\n";
+            retVal = -1;
+            continue;
+        }
+
+        std::string server = SwigWrap::GetNym_OutpaymentsNotaryIDByIndex(the_mynym, i);
+        if (server.empty()) {
+            otOut << "Error: cannot load server for payment " << i << ".\n";
+            retVal = -1;
+            continue;
+        }
+
+        // Done: Put the code here where we message the server to cancel all
+        // relevant transaction numbers for the instrument. If it's a cheque,
+        // there's only one number. But if it's a smart contract, there could
+        // be many numbers. Seems like best thing is to just activate it, but
+        // have a "rejected" flag which causes the activation to fail. (That
+        // way, all the other parties will be properly notified, which the
+        // server already does.) We don't even need to remove it from the
+        // outpayment box, because the failure notification from the server
+        // will already cause the OTClient to remove it from the outpayment box.
+        //
+        // Ah-ha! ANY outgoing payment plan or smart contract is necessarily
+        // incomplete: it's outgoing because it was sent to the next party so
+        // he could sign it, too, and probably activate it. Presumably he has
+        // not done so yet (if I am 'beating him to the punch' by cancelling it
+        // before he can activate it) and therefore the plan or smart contract
+        // still is missing at least one signer, so it is GUARANTEED to fail
+        // verification if I try to activate it myself. (Good.)
+        //
+        // This means I can just take whatever instrument appears outgoing,
+        // and try to activate it myself. It will definitely fail activation,
+        // and then the failure notice will already be sent from that, to all
+        // the parties, and they can harvest back their numbers automatically
+        // as necessary.
+        //
+        // The one problem is, though this works for payment plans and smart
+        // contracts, it will not work for cheques. The cheque is made out to
+        // someone else, and he is the one normally who needs to deposit it.
+        // Plus, I can't deposit a cheque into the same account it's drawn on.
+        //
+        // UPDATE: Now when a cheque is deposited into the same account it was
+        // drawn on, that will be interpreted by the server as a request to
+        // CANCEL the cheque.
+
+        const Identifier theNotaryID{server}, theNymID{the_mynym};
+        std::string type = SwigWrap::Instrmnt_GetType(payment);
+
+        if ("SMARTCONTRACT" == type) {
+            // Just take the smart contract from the outpayment box, and try to
+            // activate it. It WILL fail, and then the failure message will be
+            // propagated to all the other parties to the contract. (Which will
+            // result in its automatic removal from the outpayment box.)
+
+            // FIX: take myacct from smart contract instead of --myacct
+            if (the_myacct.empty()) {
+                otOut << "You MUST provide --myacct for smart contracts.\n";
+                retVal = -1;
+                continue;
+            }
+
+            // Try to activate the smart contract. (As a way of  cancelling it.)
+            // So while we expect this 'activation' to fail, it should have the
+            // desired effect of cancelling the smart contract and sending
+            // failure notices to all the parties.
+            std::unique_ptr<OTSmartContract> contract =
+                std::make_unique<OTSmartContract>();
+
+            OT_ASSERT(contract)
+
+            contract->LoadContractFromString(String(payment));
+            std::string response = OT::App()
+                                  .API()
+                                  .ServerAction()
+                                  .ActivateSmartContract(
+                                      theNymID,
+                                      theNotaryID,
+                                      Identifier(the_myacct),
+                                      "acct_agent_name",
+                                      contract)
+                                  ->Run();
+            if (response.empty()) {
+                otOut << "Error: cannot cancel smart contract.\n";
+                retVal = -1;
+                continue;
+            }
+
+            otOut << "Server reply: \n" << response << "\n";
+
+            if (1 != SwigWrap::Message_IsTransactionCanceled(
+                         server, the_mynym, the_myacct, response)) {
+                otOut << "Error: cancel smart contract failed.\n";
+                retVal = -1;
+                continue;
+            }
+
+            otOut << "Success canceling smart contract!\n";
+            continue;
+        }
+
+        if ("PAYMENT PLAN" == type) {
+            // Just take the payment plan from the outpayment box, and try to
+            // activate it. It WILL fail, and then the failure message will be
+            // propagated to the other party to the contract. (Which will result
+            // in its automatic removal from the outpayment box.)
+
+            std::unique_ptr<OTPaymentPlan> plan =
+                std::make_unique<OTPaymentPlan>();
+
+            OT_ASSERT(plan)
+
+            plan->LoadContractFromString(String(payment));
+            std::string response =
+                OT::App()
+                    .API()
+                    .ServerAction()
+                    .CancelPaymentPlan(theNymID, theNotaryID, plan)
+                    ->Run();
+            if (response.empty()) {
+                otOut << "Error: cannot cancel payment plan.\n";
+                retVal = -1;
+                continue;
+            }
+
+            otOut << "Server reply: \n" << response << "\n";
+
+            if (1 != SwigWrap::Message_IsTransactionCanceled(
+                         server, the_mynym, the_myacct, response)) {
+                otOut << "Error: cancel payment plan failed.\n";
+                retVal = -1;
+                continue;
+            }
+
+            otOut << "Success canceling payment plan!\n";
+            continue;
+        }
+
+        if ("PURSE" == type) {
+            // This is a tricky one -- why would anyone EVER want to discard
+            // outgoing cash? Normally your incentive would be to do the
+            // opposite: Keep a copy of all outgoing cash until the copy
+            // itself expires (when the cash expires.) This way it's always
+            // recoverable in the event of a "worst case" situation.
+            //
+            // So what do we do in this case? Nevertheless, the user has
+            // explicitly just instructed the client to DISCARD OUTGOING CASH.
+            //
+            // Perhaps we should just ask the user to CONFIRM that he wants to
+            // erase the cash, and make SURE that he understands the
+            // consequences of that choice.
+
+            // removes payment instrument (from payments in or out box)
+            if (!SwigWrap::RecordPayment(server, the_mynym, false, i, false)) {
+                otOut << "Error: cannot cancel cash purse.\n";
+                retVal = -1;
+                continue;
+            }
+
+            otOut << "Success canceling cash purse!\n";
+            continue;
+        }
+
+        // CHEQUE VOUCHER INVOICE
+
+        bool isVoucher = ("VOUCHER" == type);
+
+        // Get the nym and account IDs from the cheque itself.
+        std::string acctID = isVoucher
+                            ? SwigWrap::Instrmnt_GetRemitterAcctID(payment)
+                            : SwigWrap::Instrmnt_GetSenderAcctID(payment);
+        if (acctID.empty()) {
+            otOut << "Error: cannot retrieve asset account ID.\n";
+            retVal = -1;
+            continue;
+        }
+
+        std::string nymID = isVoucher ? SwigWrap::Instrmnt_GetRemitterNymID(payment)
+                                 : SwigWrap::Instrmnt_GetSenderNymID(payment);
+        if (nymID.empty()) {
+            otOut << "Error: cannot retrieve sender nym.\n";
+            retVal = -1;
+            continue;
+        }
+
+        if (nymID != the_mynym) {
+            otOut << "Error: unexpected sender nym.\n";
+            retVal = -1;
+            continue;
+        }
+
+        if (1 != depositCheque(server, acctID, nymID, payment)) {
+            otOut << "Error: cannot cancel " << type << ".\n";
+            retVal = -1;
+            continue;
+        }
+
+        otOut << "Success canceling " << type << "!\n";
+    }
+    return retVal;
+}
+
+std::int32_t OTRecordList::acceptFromInbox( // a static method
+    const std::string& myacct,
+    const std::string& indices,
+    const std::int32_t itemTypeFilter)
+{
+    std::string server = SwigWrap::GetAccountWallet_NotaryID(myacct);
+    if (server.empty()) {
+        otOut << "Error: cannot determine server from myacct.\n";
+        return -1;
+    }
+
+    std::string mynym = SwigWrap::GetAccountWallet_NymID(myacct);
+    if (mynym.empty()) {
+        otOut << "Error: cannot determine mynym from myacct.\n";
+        return -1;
+    }
+    // -----------------------------------------------------------
+    // NOTE: I just removed this during my most recent changes. It just seems,
+    // now with Justus' regular automated refreshes, I shouldn't have to grab
+    // these right BEFORE, when I then anyway have to grab them right AFTER
+    // (at the end of this function) once it succeeds. Should speed things
+    // up?
+    // Also, I considered the fact that we're using indices in this function
+    // So if the user has actually selected certain indices already, then seems
+    // unwise to download the inbox before processing THOSE indices. Rather
+    // have it fail and re-try, and at least be trying the actual intended
+    // indices, and cut our account retrievals in half while we're at it!
+    //
+    //    if (!OT::App().API().ME().retrieve_account(server, mynym, myacct,
+    //    true)) {
+    //        otOut << "Error retrieving intermediary files for account.\n";
+    //        return -1;
+    //    }
+    // -----------------------------------------------------------
+    // NOTE: Normally we don't have to do this, because the high-level API is
+    // smart enough, when sending server transaction requests, to grab new
+    // transaction numbers if it is running low. But in this case, we need the
+    // numbers available BEFORE sending the transaction request, because the
+    // call to SwigWrap::Ledger_CreateResponse is where the number is first
+    // needed, and that call is made before the server transaction request is
+    // actually sent.
+    //
+    const Identifier theNotaryID{server}, theNymID{mynym}, theAcctID{myacct};
+
+    if (!OT::App().API().ServerAction().GetTransactionNumbers(
+            theNymID, theNotaryID, 10)) {
+        otOut << "Error: cannot reserve transaction numbers.\n";
+        return -1;
+    }
+    // -----------------------------------------------------------
+    std::unique_ptr<Ledger> pInbox(
+        OT::App().API().OTAPI().LoadInbox(theNotaryID, theNymID, theAcctID));
+    if (false == bool(pInbox)) {
+        otOut << "Error: cannot load inbox.\n";
+        return -1;
+    }
+    // -----------------------------------------------------------
+    std::int32_t item_count = pInbox->GetTransactionCount();
+    // -----------------------------------------------------------
+    if (0 > item_count) {
+        otErr << "Error: cannot load inbox item count.\n";
+        return -1;
+    } else if (0 == item_count) {
+        otWarn << "The inbox is empty.\n";
+        return 0;
+    }
+
+    if (!OTRecordList::checkIndicesRange("indices", indices, item_count)) {
+        return -1;
+    }
+
+    bool all = ("" == indices || "all" == indices);
+    // -----------------------------------------------------------
+    std::set<int32_t>* pOnlyForIndices{nullptr};
+    std::set<int32_t> setForIndices;
+    if (!all) {
+        NumList numlistForIndices{indices};
+        std::set<int64_t> setForIndices64;
+        if (numlistForIndices.Output(setForIndices64)) {
+            pOnlyForIndices = &setForIndices;
+            for (const int64_t& lIndex : setForIndices64) {
+                setForIndices.insert(static_cast<int32_t>(lIndex));
+            }
+        }
+    }
+    std::set<int64_t> receiptIds{pInbox->GetTransactionNums(pOnlyForIndices)};
+
+    if (receiptIds.size() < 1) {
+        otWarn << "There are no inbox receipts to process.\n";
+        return 0;
+    }
+    // -----------------------------------------------------------
+    // NOTE: Indices are only optional. Otherwise it's "accept all receipts".
+    // But even if you DID pass in a comma-separated list of indices, it doesn't
+    // matter below this point.
+    // That's because we have now translated the user-selected GUI indices into
+    // an actual set of receipt IDs, each being the trans num on a transaction
+    // inside the inbox.
+
+    // -------------------------------------------------------
+    OT_API::ProcessInbox response{OT::App().API().OTAPI().Ledger_CreateResponse(
+        theNotaryID, theNymID, theAcctID)};
+    // -------------------------------------------------------
+    auto& processInbox = std::get<0>(response);
+    auto& inbox = std::get<1>(response);
+
+    if (!bool(processInbox) || !bool(inbox)) {
+        otWarn << __FUNCTION__ << "Ledger_CreateResponse somehow failed.\n";
+        return -1;
+    }
+    // -------------------------------------------------------
+    for (const int64_t& lReceiptId : receiptIds) {
+        OTTransaction* pReceipt =
+            OT::App().API().OTAPI().Ledger_GetTransactionByID(
+                *inbox, lReceiptId);
+
+        if (nullptr == pReceipt) {
+            otErr << __FUNCTION__
+                  << "Unexpectedly got a nullptr for ReceiptId: " << lReceiptId;
+            return -1;
+        }  // Below this point, pReceipt is a good pointer. It's
+        //   owned by inbox, so no need to delete.
+        // ------------------------
+        // itemTypeFilter == 0 for all, 1 for transfers only, 2 for receipts
+        // only.
+        //
+        if (0 != itemTypeFilter) {
+            const OTTransaction::transactionType receipt_type{
+                pReceipt->GetType()};
+
+            const bool transfer = (OTTransaction::pending == receipt_type);
+
+            if ((1 == itemTypeFilter) && !transfer) {
+                // not a pending transfer.
+                continue;
+            }
+            if ((2 == itemTypeFilter) && transfer) {
+                // not a receipt.
+                continue;
+            }
+        }
+        // ------------------------
+        const bool bReceiptResponseCreated =
+            OT::App().API().OTAPI().Transaction_CreateResponse(
+                theNotaryID,
+                theNymID,
+                theAcctID,
+                *processInbox,
+                *pReceipt,
+                true);
+
+        if (!bReceiptResponseCreated) {
+            otErr << __FUNCTION__
+                  << "Error: cannot create transaction response.\n";
+            return -1;
+        }
+    }  // for
+    // -------------------------------------------------------
+    if (processInbox->GetTransactionCount() <= 0) {
+        // did not process anything
+        otErr << __FUNCTION__
+              << "Should never happen. Might want to follow up if you see this "
+                 "log.\n";
+        return 0;
+    }
+    // ----------------------------------------------
+    const bool bFinalized = OT::App().API().OTAPI().Ledger_FinalizeResponse(
+        theNotaryID, theNymID, theAcctID, *processInbox);
+
+    if (!bFinalized) {
+        otErr << __FUNCTION__ << "Error: cannot finalize response.\n";
+        return -1;
+    }
+    // ----------------------------------------------
+    const std::string notary_response =
+        OT::App()
+            .API()
+            .ServerAction()
+            .ProcessInbox(theNymID, theNotaryID, theAcctID, processInbox)
+            ->Run();
+    std::int32_t reply =
+        InterpretTransactionMsgReply(server, mynym, myacct, "process_inbox",
+                                     notary_response);
+
+    if (1 != reply) {
+        return reply;
+    }
+
+    // We KNOW they all just changed, since we just processed
+    // the inbox. Might as well refresh our copy with the new changes.
+    //
+    if (!OT::App().API().ServerAction().DownloadAccount(
+            theNymID, theNotaryID, theAcctID, true)) {
+        otOut << __FUNCTION__
+              << "Success processing inbox, but then failed "
+                 "retrieving intermediary files for account.\n";
+//      return -1; // By this point we DID successfully process the
+//      inbox.
+        // (We just then subsequently failed to download the updated acct
+        // files.)
+    }
+
+    return 1;
+}
+
+
+
+
+
+
 
 void OTRecordList::AddAccountID(std::string str_id)
 {
