+++ src/client/commands/CmdAcceptInbox.cpp
@@ -43,6 +43,8 @@
 #include "opentxs/client/commands/CmdBase.hpp"
 #include "opentxs/client/commands/CmdBaseAccept.hpp"
 
+#include "opentxs/client/OTRecordList.hpp"
+
 #include <stdint.h>
 #include <string>
 
@@ -76,5 +78,5 @@ int32_t CmdAcceptInbox::run(string myacct, string indices)
         return -1;
     }
 
-    return acceptFromInbox(myacct, indices, typeBoth);
+    return acceptFromInbox(myacct, indices, OTRecordList::typeBoth);
 }
