+++ src/client/commands/CmdBase.cpp
@@ -89,35 +89,7 @@ CmdBase::~CmdBase() {}
 
 bool CmdBase::checkAccount(const char* name, string& account) const
 {
-    if (!checkMandatory(name, account)) {
-        return false;
-    }
-
-    std::shared_ptr<Account> pAccount{nullptr};
-    OTWallet* wallet = getWallet();
-
-    Identifier theID(account);
-
-    if (!theID.empty()) pAccount = wallet->GetAccount(theID);
-
-    if (false == bool(pAccount)) {
-        pAccount = wallet->GetAccountPartialMatch(account);
-
-        if (false == bool(pAccount)) {
-            otOut << "Error: " << name << ": unknown account: " << account
-                  << "\n";
-            return false;
-        }
-    }
-
-    if (pAccount) {
-        String tmp;
-        pAccount->GetPurportedAccountID().GetString(tmp);
-        account = tmp.Get();
-    }
-
-    otWarn << "Using " << name << ": " << account << "\n";
-    return true;
+    return OTRecordList::checkAccount(name, account);
 }
 
 int64_t CmdBase::checkAmount(
@@ -175,62 +147,17 @@ int32_t CmdBase::checkIndex(
 
 bool CmdBase::checkIndices(const char* name, const string& indices) const
 {
-    if (!checkMandatory(name, indices)) {
-        return false;
-    }
-
-    if ("all" == indices) {
-        return true;
-    }
-
-    for (string::size_type i = 0; i < indices.length(); i++) {
-        if (!isdigit(indices[i])) {
-            otOut << "Error: " << name << ": not a value: " << indices << "\n";
-            return false;
-        }
-        for (i++; i < indices.length() && isdigit(indices[i]); i++) {
-        }
-        if (i < indices.length() && ',' != indices[i]) {
-            otOut << "Error: " << name << ": not a value: " << indices << "\n";
-            return false;
-        }
-    }
-
-    return true;
+    return OTRecordList::checkIndices(name, indices);
 }
 
 bool CmdBase::checkMandatory(const char* name, const string& value) const
 {
-    if ("" == value) {
-        otOut << "Error: " << name << ": mandatory parameter not specified.\n";
-        return false;
-    }
-
-    return true;
+    return OTRecordList::checkMandatory(name, value);
 }
 
 bool CmdBase::checkNym(const char* name, string& nym, bool checkExistance) const
 {
-    if (!checkMandatory(name, nym)) return false;
-
-    ConstNym pNym = nullptr;
-    const Identifier nymID(nym);
-
-    if (!nymID.empty()) pNym = OT::App().Wallet().Nym(nymID);
-
-    if (nullptr == pNym) pNym = OT::App().Wallet().NymByIDPartialMatch(nym);
-
-    if (nullptr != pNym) {
-        String tmp;
-        pNym->GetIdentifier(tmp);
-        nym = tmp.Get();
-    } else if (checkExistance) {
-        otOut << "Error: " << name << ": unknown nym: " << nym << "\n";
-        return false;
-    }
-
-    otOut << "Using " << name << ": " << nym << "\n";
-    return true;
+    return OTRecordList::checkNym(name, nym, checkExistance);
 }
 
 bool CmdBase::checkPurse(const char* name, string& purse) const
@@ -289,54 +216,7 @@ bool CmdBase::checkPurse(const char* name, string& purse) const
 
 bool CmdBase::checkServer(const char* name, string& server) const
 {
-    if (!checkMandatory(name, server)) return false;
-
-    Identifier theID(server);
-    ConstServerContract pServer;  // shared_ptr to const.
-
-    // See if it's available using the full length ID.
-    if (!theID.empty()) pServer = OT::App().Wallet().Server(theID);
-
-    if (!pServer) {
-        const auto servers = OT::App().Wallet().ServerList();
-
-        // See if it's available using the partial length ID.
-        for (auto& it : servers) {
-            if (0 == it.first.compare(0, server.length(), server)) {
-                pServer = OT::App().Wallet().Server(Identifier(it.first));
-                break;
-            }
-        }
-        if (!pServer) {
-            // See if it's available using the full length name.
-            for (auto& it : servers) {
-                if (0 == it.second.compare(0, it.second.length(), server)) {
-                    pServer = OT::App().Wallet().Server(Identifier(it.first));
-                    break;
-                }
-            }
-
-            if (!pServer) {
-                // See if it's available using the partial name.
-                for (auto& it : servers) {
-                    if (0 == it.second.compare(0, server.length(), server)) {
-                        pServer =
-                            OT::App().Wallet().Server(Identifier(it.first));
-                        break;
-                    }
-                }
-            }
-        }
-    }
-
-    if (!pServer) {
-        otOut << "Error: " << name << ": unknown server: " << server << "\n";
-        return false;
-    }
-
-    server = pServer->ID().str();
-    otOut << "Using " << name << ": " << server << "\n";
-    return true;
+    return OTRecordList::checkServer(name, server);
 }
 
 int64_t CmdBase::checkTransNum(const char* name, const string& id) const
